WIP: decimal arithmetic
