+++ polars/polars-arrow/src/compute/arithmetics/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "dtype-decimal")]
+pub mod decimal;
